+++ relm4/src/actions/mod.rs
@@ -1,8 +1,7 @@
 //! Action utility.
 
 use gtk::gio;
-use gtk::glib::FromVariant;
-use gtk::prelude::{ActionExt, ActionMapExt, StaticVariantType, ToVariant, WidgetExt};
+use gtk::prelude::{ActionExt, ActionMapExt, FromVariant, StaticVariantType, ToVariant, WidgetExt};
 
 use std::marker::PhantomData;